requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email_validator
python-dotenv
psycopg2-binary
Flask-Migrate